Function hacspec_lib::prelude::cmul

pub fn cmul<T: Integer + Copy>(x: T, y: T, c: T) -> T

Function that is not part of the language but is offered as a helper for tests, etc. Multiply two numerics if condition c is set (all bits 1). Returns x if condition c is 0. Note: Multiplication is always wrapping.