Struct hacspec_lib::prelude::I8

pub struct I8(pub i8);

Tuple Fields
0: i8
Implementations

impl I8

pub fn classify<T>(x: T) -> I8where
T: Into<i8>,

pub fn declassify(self) -> i8

Warning: use with caution, breaks the constant-time guarantee.

pub fn zero() -> I8

pub fn one() -> I8

pub fn ones() -> I8

pub fn from_le_bytes(bytes: &[U8]) -> Vec<I8, Global>

pub fn to_le_bytes(ints: &[I8]) -> Vec<U8, Global>

pub fn from_be_bytes(bytes: &[U8]) -> Vec<I8, Global>

pub fn to_be_bytes(ints: &[I8]) -> Vec<U8, Global>

pub fn max_value() -> I8

impl I8

pub fn checked_add(self, rhs: I8) -> I8

Warning: panics when overflow.

impl I8

pub fn checked_sub(self, rhs: I8) -> I8

Warning: panics when overflow.

impl I8

pub fn checked_mul(self, rhs: I8) -> I8

Warning: panics when overflow.

impl I8

pub fn rotate_left(self, rotval: usize) -> I8

pub fn rotate_right(self, rotval: usize) -> I8

impl I8

Constant-time comparison operators

pub fn comp_eq(self, rhs: I8) -> I8

pub fn comp_ne(self, rhs: I8) -> I8

Produces a new integer which is all ones if the first argument is different from the second argument, and all zeroes otherwise.

pub fn comp_gte(self, rhs: I8) -> I8

Produces a new integer which is all ones if the first argument is greater than or equal to the second argument, and all zeroes otherwise. With inspiration from

pub fn comp_gt(self, rhs: I8) -> I8

Produces a new integer which is all ones if the first argument is strictly greater than the second argument, and all zeroes otherwise.

pub fn comp_lte(self, rhs: I8) -> I8

Produces a new integer which is all ones if the first argument is less than or equal to the second argument, and all zeroes otherwise.

pub fn comp_lt(self, rhs: I8) -> I8

Produces a new integer which is all ones if the first argument is strictly less than the second argument, and all zeroes otherwise.