Struct hacspec_lib::prelude::U32

pub struct U32(pub u32);

Tuple Fields
0: u32
Implementations

impl U32

Constant-time comparison operators

pub fn comp_eq(self, rhs: U32) -> U32

Produces a new integer which is all ones if the two arguments are equal and all zeroes otherwise. With inspiration from Wireguard.

pub fn comp_ne(self, rhs: U32) -> U32

Produces a new integer which is all ones if the first argument is different from the second argument, and all zeroes otherwise.

pub fn comp_gte(self, rhs: U32) -> U32

Produces a new integer which is all ones if the first argument is greater than or equal to the second argument, and all zeroes otherwise. With inspiration from WireGuard.

pub fn comp_gt(self, rhs: U32) -> U32

Produces a new integer which is all ones if the first argument is strictly greater than the second argument, and all zeroes otherwise.

pub fn comp_lte(self, rhs: U32) -> U32

Produces a new integer which is all ones if the first argument is less than or equal to the second argument, and all zeroes otherwise.

pub fn comp_lt(self, rhs: U32) -> U32

Produces a new integer which is all ones if the first argument is strictly less than the second argument, and all zeroes otherwise.

impl U32

pub fn classify<T>(x: T) -> U32 where
T: Into<u32>,

pub fn declassify(self) -> u32

Warning: use with caution, breaks the constant-time guarantee.

pub fn zero() -> U32

pub fn one() -> U32

pub fn ones() -> U32

pub fn from_le_bytes(bytes: &[U8]) -> Vec<U32, Global>

pub fn to_le_bytes(ints: &[U32]) -> Vec<U8, Global>

pub fn from_be_bytes(bytes: &[U8]) -> Vec<U32, Global>

pub fn to_be_bytes(ints: &[U32]) -> Vec<U8, Global>

pub fn max_value() -> U32

impl U32

pub fn rotate_left(self, rotval: usize) -> U32

pub fn rotate_right(self, rotval: usize) -> U32

impl U32

pub fn checked_add(self, rhs: U32) -> U32

Warning: panics when overflow.

impl U32

pub fn checked_sub(self, rhs: U32) -> U32

Warning: panics when overflow.

impl U32

pub fn checked_mul(self, rhs: U32) -> U32

Warning: panics when overflow.