Struct hacspec_lib::prelude::U8

pub struct U8(pub u8);

Implementations

impl U8

Constant-time comparison operators

pub fn comp_eq(self, rhs: U8) -> U8

Produces a new integer which is all ones if the two arguments are equal and all zeroes otherwise. With inspiration from Wireguard.

pub fn comp_ne(self, rhs: U8) -> U8

Produces a new integer which is all ones if the first argument is different from the second argument, and all zeroes otherwise.

pub fn comp_gte(self, rhs: U8) -> U8

Produces a new integer which is all ones if the first argument is greater than or equal to the second argument, and all zeroes otherwise. With inspiration from WireGuard.

pub fn comp_gt(self, rhs: U8) -> U8

Produces a new integer which is all ones if the first argument is strictly greater than the second argument, and all zeroes otherwise.

pub fn comp_lte(self, rhs: U8) -> U8

Produces a new integer which is all ones if the first argument is less than or equal to the second argument, and all zeroes otherwise.

pub fn comp_lt(self, rhs: U8) -> U8

Produces a new integer which is all ones if the first argument is strictly less than the second argument, and all zeroes otherwise.

impl U8

pub fn classify<T>(x: T) -> U8 where
T: Into<u8>,

pub fn declassify(self) -> u8

Warning: use with caution, breaks the constant-time guarantee.

pub fn zero() -> U8

pub fn one() -> U8

pub fn ones() -> U8

pub fn from_le_bytes(bytes: &[U8]) -> Vec<U8>

pub fn to_le_bytes(ints: &[U8]) -> Vec<U8>

pub fn from_be_bytes(bytes: &[U8]) -> Vec<U8>

pub fn to_be_bytes(ints: &[U8]) -> Vec<U8>

pub fn max_value() -> U8

impl U8

pub fn rotate_left(self, rotval: usize) -> U8

pub fn rotate_right(self, rotval: usize) -> U8

impl U8

pub fn checked_add(self, rhs: U8) -> U8

Warning: panics when overflow.

impl U8

pub fn checked_sub(self, rhs: U8) -> U8

Warning: panics when overflow.

impl U8

pub fn checked_mul(self, rhs: U8) -> U8

Warning: panics when overflow.