Struct hacspec_lib::prelude::u32Word

pub struct u32Word(pub [u8; 4]);

Fixed length byte array.

Tuple Fields
0: [u8; 4]
Implementations

impl u32Word

pub fn new() -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn length() -> usize

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn from_array(v: [u8; 4]) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn from_native_slice(v: &[u8]) -> Self

Function that is not part of the language but is offered as a helper for tests, etc.

impl u32Word

pub fn from_slice<A: SeqTrait<u8>>(input: &A, start: usize, len: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn concat<A: SeqTrait<u8>>(&self, next: &A) -> Seq<u8>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn from_slice_range<A: SeqTrait<u8>>(input: &A, r: Range<usize>) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn slice(&self, start_out: usize, len: usize) -> Seq<u8>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn slice_range(&self, r: Range<usize>) -> Seq<u8>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn num_chunks(&self, chunk_size: usize) -> usize

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn get_chunk_len(&self, chunk_size: usize, chunk_number: usize) -> usize

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn get_chunk(
&self,
chunk_size: usize,
chunk_number: usize
) -> (usize, Seq<u8>)

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn set_chunk<A: SeqTrait<u8>>(
self,
chunk_size: usize,
chunk_number: usize,
input: &A
) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl u32Word

pub fn from_vec(x: Vec<u8>) -> u32Word

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_seq<T: SeqTrait<u8>>(x: &T) -> u32Word

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl u32Word

pub fn from_hex(s: &str) -> u32Word

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec Read hex string to Bytes.

impl u32Word

pub fn to_be_u32s(&self) -> Seq<u32>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_le_u32s(&self) -> Seq<u32>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_be_u64s(&self) -> Seq<u64>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_le_u64s(&self) -> Seq<u64>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_u128s_be(&self) -> Seq<u128>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_u128s_le(&self) -> Seq<u128>

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn to_hex(&self) -> String

Function that is not part of the language but is offered as a helper for tests, etc.