Trait hacspec_lib::prelude::Integer

pub trait Integer: Numeric {
    const NUM_BITS: usize;

    fn ZERO() -> Self;
    fn ONE() -> Self;
    fn TWO() -> Self;
    fn from_literal(val: u128) -> Self;
    fn from_hex_string(s: &String) -> Self;
    fn get_bit(self, i: usize) -> Self;
    fn set_bit(self, b: Self, i: usize) -> Self;
    fn set(self, pos: usize, y: Self, yi: usize) -> Self;
    fn rotate_left(self, n: usize) -> Self;
    fn rotate_right(self, n: usize) -> Self;
}

This trait extends the Numeric trait and is implemented by all integer types. It offers bit manipulation, instantiation from literal, and convenient constants.

Required Associated Constants

const NUM_BITS: usize

Required Methods

fn ZERO() -> Self

fn ONE() -> Self

fn TWO() -> Self

fn from_literal(val: u128) -> Self

Get an integer with value val.

fn from_hex_string(s: &String) -> Self

Read a hex string (starting with 0x) into an Integer.

fn get_bit(self, i: usize) -> Self

fn set_bit(self, b: Self, i: usize) -> Self

fn set(self, pos: usize, y: Self, yi: usize) -> Self

fn rotate_left(self, n: usize) -> Self

fn rotate_right(self, n: usize) -> Self