Trait hacspec_lib::prelude::UnsignedSecretInteger

pub trait UnsignedSecretInteger: UnsignedInteger + SecretInteger {
    fn to_le_bytes(self) -> Seq<U8>;

    fn to_be_bytes(self) -> Seq<U8>;

    fn from_le_bytes(x: &Seq<U8>) -> Self;

    fn from_be_bytes(x: &Seq<U8>) -> Self;

    fn get_byte(self, i: usize) -> Self { ... }
}

Required methods

fn to_le_bytes(self) -> Seq<U8>

fn to_be_bytes(self) -> Seq<U8>

fn from_le_bytes(x: &Seq<U8>) -> Self

fn from_be_bytes(x: &Seq<U8>) -> Self

Loading content...
Provided methods

fn get_byte(self, i: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

Get byte i of this integer.

Loading content...