Struct hacspec_lib::seq::PublicSeq

pub struct PublicSeq<T: Default + PublicInteger> { /* private fields */ }

Variable length byte arrays.

Implementations

impl PublicSeq<u8>

pub fn from_public_slice(v: &[u8]) -> PublicSeq<u8>

pub fn from_native(b: Vec<u8>) -> PublicSeq<u8>

pub fn to_native(&self) -> Vec<u8>

pub fn into_native(self) -> Vec<u8>

pub fn as_slice(&self) -> &[u8]

impl<T: Clone + Default + PublicInteger> PublicSeq<T>

pub fn new(l: usize) -> Self

pub fn with_capacity(l: usize) -> Self

pub fn reserve(self, additional: usize) -> Self

pub fn len(&self) -> usize

Get the size of this sequence.

pub fn slice(&self, start_out: usize, len: usize) -> Self

pub fn native_slice(&self) -> &[T]

pub fn into_slice(self, start_out: usize, len: usize) -> Self

pub fn slice_range(&self, r: Range<usize>) -> Self

pub fn into_slice_range(self, r: Range<usize>) -> Self

pub fn split_off(self, at: usize) -> (Self, Self)

pub fn pop(self) -> (T, Self)

pub fn truncate(self, len: usize) -> Self

pub fn from_slice<A: SeqTrait<T>>(input: &A, start: usize, len: usize) -> Self

pub fn concat<A: SeqTrait<T>>(&self, next: &A) -> Self

pub fn concat_owned(self, next: Self) -> Self

pub fn push(&self, next: &T) -> Self

pub fn push_owned(self, next: T) -> Self

pub fn from_slice_range<A: SeqTrait<T>>(input: &A, r: Range<usize>) -> Self

pub fn num_chunks(&self, chunk_size: usize) -> usize

pub fn num_exact_chunks(&self, chunk_size: usize) -> usize

Get the number of chunks of chunk_size in this array. There might be less than chunk_size remaining elements in this array beyond these.

pub fn get_chunk(&self, chunk_size: usize, chunk_number: usize) -> (usize, Self)

pub fn get_exact_chunk(&self, chunk_size: usize, chunk_number: usize) -> Self

Get the chunk_number chunk of chunk_size from this array as Seq<T>. The resulting sequence is of exactly chunk_size length. Until #84 is fixed this returns an empty sequence if not enough elements are left.

pub fn get_remainder_chunk(&self, chunk_size: usize) -> Self

Get the remaining chunk of this array of length less than chunk_size. If there’s no remainder, i.e. if the length of this array can be divided by chunk_size without a remainder, the function returns an empty sequence (until #84 is fixed).

pub fn set_chunk<A: SeqTrait<T>>(
self,
chunk_size: usize,
chunk_number: usize,
input: &A
) -> Self

pub fn set_exact_chunk<A: SeqTrait<T>>(
self,
chunk_size: usize,
chunk_number: usize,
input: &A
) -> Self

pub fn update_owned(self, start_out: usize, v: Self) -> Self

impl<T: Clone + Default + PublicInteger> PublicSeq<T>

pub fn from_vec(b: Vec<T>) -> PublicSeq<T>

pub fn from_native_slice(x: &[T]) -> PublicSeq<T>

pub fn from_seq<U: SeqTrait<T>>(x: &U) -> PublicSeq<T>

impl PublicSeq<u8>

pub fn from_hex(s: &str) -> PublicSeq<u8>

pub fn from_string(s: String) -> PublicSeq<u8>

impl PublicSeq<u8>

pub fn to_hex(&self) -> String