Struct hacspec_lib::seq::Seq

pub struct Seq<T: Default> { /* fields omitted */ }

Variable length byte arrays.

Implementations

impl<T: Clone + Default> Seq<T>

pub fn new(l: usize) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn with_capacity(l: usize) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn reserve(self, additional: usize) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn len(&self) -> usize

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

Get the size of this sequence.

pub fn slice(&self, start_out: usize, len: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn native_slice(&self) -> &[T]

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn into_slice(self, start_out: usize, len: usize) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn slice_range(&self, r: Range<usize>) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_slice_range(self, r: Range<usize>) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn split_off(self, at: usize) -> (Self, Self)

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn truncate(self, len: usize) -> Self

This function can be called from hacspec programs but its body features Rust constructs that are not part of hacspec

pub fn from_slice<A: SeqTrait<T>>(input: &A, start: usize, len: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn concat<A: SeqTrait<T>>(&self, next: &A) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn concat_owned(self, next: Self) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn push(&self, next: &T) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn push_owned(self, next: T) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn from_slice_range<A: SeqTrait<T>>(input: &A, r: Range<usize>) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn num_chunks(&self, chunk_size: usize) -> usize

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn num_exact_chunks(&self, chunk_size: usize) -> usize

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

Get the number of chunks of chunk_size in this array. There might be less than chunk_size remaining elements in this array beyond these.

pub fn get_chunk(&self, chunk_size: usize, chunk_number: usize) -> (usize, Self)

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn get_exact_chunk(&self, chunk_size: usize, chunk_number: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

Get the chunk_number chunk of chunk_size from this array as Seq<T>. The resulting sequence is of exactly chunk_size length. Until #84 is fixed this returns an empty sequence if not enough elements are left.

pub fn get_remainder_chunk(&self, chunk_size: usize) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

Get the remaining chunk of this array of length less than chunk_size. If there's no remainder, i.e. if the length of this array can be divided by chunk_size without a remainder, the function returns an empty sequence (until #84 is fixed).

pub fn set_chunk<A: SeqTrait<T>>(
self,
chunk_size: usize,
chunk_number: usize,
input: &A
) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn set_exact_chunk<A: SeqTrait<T>>(
self,
chunk_size: usize,
chunk_number: usize,
input: &A
) -> Self

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl<T: Clone + Default> Seq<T>

pub fn from_vec(b: Vec<T>) -> Seq<T>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_native_slice(x: &[T]) -> Seq<T>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_seq<U: SeqTrait<T>>(x: &U) -> Seq<T>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U8>

Read hex string to Bytes.

pub fn from_hex(s: &str) -> Seq<U8>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_string(s: String) -> Seq<U8>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U8>

pub fn from_public_slice(v: &[u8]) -> Seq<U8>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_public_seq<U: SeqTrait<u8>>(x: &U) -> Seq<U8>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U16>

pub fn from_public_slice(v: &[u16]) -> Seq<U16>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_public_seq<U: SeqTrait<u16>>(x: &U) -> Seq<U16>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U32>

pub fn from_public_slice(v: &[u32]) -> Seq<U32>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_public_seq<U: SeqTrait<u32>>(x: &U) -> Seq<U32>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U64>

pub fn from_public_slice(v: &[u64]) -> Seq<U64>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_public_seq<U: SeqTrait<u64>>(x: &U) -> Seq<U64>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U128>

pub fn from_public_slice(v: &[u128]) -> Seq<U128>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn from_public_seq<U: SeqTrait<u128>>(x: &U) -> Seq<U128>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

impl Seq<U8>

pub fn declassify(self) -> Seq<u8>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_native(self) -> Vec<u8>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn to_native(&self) -> Vec<u8>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U16>

pub fn declassify(self) -> Seq<u16>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_native(self) -> Vec<u16>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn to_native(&self) -> Vec<u16>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U32>

pub fn declassify(self) -> Seq<u32>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_native(self) -> Vec<u32>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn to_native(&self) -> Vec<u32>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U64>

pub fn declassify(self) -> Seq<u64>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_native(self) -> Vec<u64>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn to_native(&self) -> Vec<u64>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U128>

pub fn declassify(self) -> Seq<u128>

This function is within the hacspec subset of Rust: its signature and body use only hacspec constructs and call functions whose signatures are in hacspec.

pub fn into_native(self) -> Vec<u128>

Function that is not part of the language but is offered as a helper for tests, etc.

pub fn to_native(&self) -> Vec<u128>

Function that is not part of the language but is offered as a helper for tests, etc.

impl Seq<U8>

pub fn to_hex(&self) -> String

Function that is not part of the language but is offered as a helper for tests, etc.