Module hacspec_lib::array

Arrays

This module implements fixed-length arrays and utility functions for it.

Note that all macros starting with an underscore (_array_base etc.) are note intended for public use. Unfortunately it’s not possible to hide them.

To define a new array type with name State, holding 16 u32 run

use hacspec_lib::*;
array!(State, 16, u32, type_for_indexes: StateIdx);

Instantiating Arrays

There are several different ways of creating array types.