Trait hacspec_lib::prelude::UnsignedSecretInteger

pub trait UnsignedSecretInteger: UnsignedInteger + SecretInteger {
    fn to_le_bytes(self) -> Seq<U8>;

    fn to_be_bytes(self) -> Seq<U8>;

    fn from_le_bytes(x: &Seq<U8>) -> Self;

    fn from_be_bytes(x: &Seq<U8>) -> Self;

    fn get_byte(self, i: usize) -> Self { ... }
}

Required methods

fn to_le_bytes(self) -> Seq<U8>

fn to_be_bytes(self) -> Seq<U8>

fn from_le_bytes(x: &Seq<U8>) -> Self

fn from_be_bytes(x: &Seq<U8>) -> Self

Loading content...
Provided methods

fn get_byte(self, i: usize) -> Self

Library function that is part of the language but implemented only with primitives and other library functions. As such, it does not need to be formalized as a primitive.

Get byte i of this integer.

Loading content...